refactor: move loader to new fiel
 - barebones TestResult class
 - removed debugging statements
